Bug fix: Docker container initialization
Fixes an issue where Docker containers associated with an IPTB node would be
ignored on `iptb start` and a new container would be created instead. Now the
existing container starts instead. This is primarily a concern when multiple
`iptb start`s are run consecutively (with no `iptb kill` in between), which
happens e.g. when the first `iptb start` fails after writing to the `dockerID`
file.

License: MIT
Signed-off-by: David Grisham <dgrisham@mines.edu>