Kill Docker node on failed start.
Whenever the `Start()` function for a Docker node is called, there's the chance
that an error occurs after the container has been started. Previously, the
container would continue to run even after the error occurred -- this is likely
confusing from the user's perspective, because they don't expect to receive an
error on initialization but still have a container running after. The way to fix
fix this in the past would've been to run `iptb stop` before trying `ipfs start`
again. This commit makes it so that the container is killed once an error occurs
to prevent this confusion.

License: MIT
Signed-off-by: David Grisham <dgrisham@mines.edu>