Can now input file with list of commands that are run async on nodes.
Usage: Pass `--cmdFile <cmds>` to `iptb run`, where `<cmds>` is a file where
each line is of the form `<nodeRange>: <cmd>` -- `<cmd>` is the command to
run on all of the nodes in `<nodeRange>`. The format of `<nodeRange>` should be
anything accepted by the `commands.parseRange()` function.

License: MIT
Signed-off-by: David Grisham <dgrisham@mines.edu>

rebase me